+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/StudentRestController.java
@@ -1,6 +1,8 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.StudentDTO;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.CourseDoNotExistException;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.StudentAlreadyExistsException;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.StudentController;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Course;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Student;
@@ -18,17 +20,17 @@ public class StudentRestController {
     }
 
     @PostMapping("{courseId}")
-    public Student createStudent(@RequestBody StudentDTO studentDTO, @PathVariable Long courseId) {
+    public Student createStudent(@RequestBody StudentDTO studentDTO, @PathVariable Long courseId) throws StudentAlreadyExistsException, CourseDoNotExistException {
         return studentController.createStudent(studentDTO, courseId);
     }
 
     @PostMapping("")
-    public Student createStudent(@RequestBody StudentDTO studentDTO) {
+    public Student createStudent(@RequestBody StudentDTO studentDTO) throws StudentAlreadyExistsException {
         return studentController.createStudent(studentDTO);
     }
 
     @GetMapping("/course/{courseId}")
-    public List<Student> getStudentsByCourseId(@PathVariable Long courseId) {
+    public List<Student> getStudentsByCourseId(@PathVariable Long courseId) throws CourseDoNotExistException {
         return studentController.getStudentByCourseId(courseId);
     }