+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/exceptions/StudentAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions;
+
+public class StudentAlreadyExistsException extends Exception{
+    public StudentAlreadyExistsException(String message) {
+        super(message);
+    }
+}